added s3 file storage

+++ src/modules/files/service.ts
@@ -1,8 +1,23 @@
 import { UploadedFile } from 'express-fileupload';
 import { v4 as generateId } from 'uuid';
 import { extension } from 'mime-types';
+import { S3Client, PutObjectCommand, DeleteObjectCommand } from "@aws-sdk/client-s3";
+import { createReadStream } from 'fs';
 
 import { prisma } from '../../tools/prismaClient';
+
+const s3Bucket = process.env.S3_BUCKET;
+const region = process.env.S3_REGION;
+const endpoint = process.env.S3_ENDPOINT;
+const s3AccessKey = process.env.S3_ACCESS_KEY;
+const s3SecretKey = process.env.S3_SECRET_KEY;
+
+const credentials = {
+    accessKeyId: String(s3AccessKey),
+    secretAccessKey: String(s3SecretKey)
+}
+
+const client = new S3Client({ region, endpoint, credentials });
 
 export class FileService {
 
@@ -11,7 +26,26 @@ export class FileService {
             return undefined;
         }
 
-        // TODO: delete file from storage
+        const file = await prisma.file.findUnique({
+            where: {
+                id
+            }
+        });
+
+        if (!file) {
+            return undefined;
+        }
+
+        const deleteResponse = await client.send(new DeleteObjectCommand({
+            Bucket: s3Bucket,
+            Key: `${file.hash}.${file.ext}`,
+        }));
+
+        const deleteStatus = Number(deleteResponse.$metadata.httpStatusCode);
+
+        if (!(deleteStatus >= 200 && deleteStatus < 300)) {
+            return undefined;
+        }
 
         const deletedFile = await prisma.file.delete({
             where: {
@@ -28,20 +62,33 @@ export class FileService {
         const fileInfo = JSON.parse(info);
 
         if (!fileExtension) {
-            return false;
+            return undefined;
         }
 
-        // TODO: save file to storage
+        const fileStream = createReadStream(data.tempFilePath);
+
+        const putResponse = await client.send(new PutObjectCommand({
+            Bucket: s3Bucket,
+            Key: `${fileHash}.${fileExtension}`,
+            Body: fileStream
+        }));
+
+        const putStatus = Number(putResponse.$metadata.httpStatusCode);
+
+        if (!(putStatus >= 200 && putStatus < 300)) {
+            return undefined;
+        }
 
         const file = await prisma.file.create({
             data: {
                 name: data.name,
                 size: data.size,
                 ext: fileExtension,
+                hash: fileHash,
                 mime: data.mimetype,
                 description: fileInfo.description,
                 caption: fileInfo.caption,
-                url: `/public/${fileHash}.${fileExtension}`,
+                url: `${endpoint}/${s3Bucket}/${fileHash}.${fileExtension}`,
                 creatorId: createdByUserId,
             }
         })
